backend/cloud_functions/intake_scripts_service/requirements.txt
functions-framework
google-cloud-firestore
Flask


